Completed laravel 4 hour crash course

+++ laragigs/app/Http/Controllers/ListingController.php
@@ -47,6 +47,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -61,6 +63,11 @@ class ListingController extends Controller
     // Store listing update
     public function update(Request $request, Listing $listing)
     {
+        // logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorised action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -83,7 +90,18 @@ class ListingController extends Controller
     // Delete listing
     public function destroy(Listing $listing)
     {
+        // logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorised action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    // Manage listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
